feat: move files and add aip validation
