+++ src/renderer/diff.spec.ts
