+++ src/renderer/fetch-IR-strategy.ts
