+++ src/renderer/snapshot-types.ts
